backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
# Ultralytics with Python 3.12 compatible
ultralytics==8.1.0
numpy>=1.26,<3
python-dotenv==1.0.1
# TensorFlow for soil classifier
tensorflow>=2.13,<3
torch>=2.0
